recipes

+++ assets/js/alert.js
@@ -1,9 +1,26 @@
-$('#add_user').submit(function(event) {
+$('#register').submit(function(event) {
 	alert('Data Inserted Successfully!');
 });
 
 $('#update_user').submit(function(event) {
-	alert('Update done successfully !');
+	event.preventDefault();
+
+	var unindexed_array = $(this).serializeArray();
+	var data = {};
+
+	$.map(unindexed_array, function(n, i) {
+		data[n['name']] = n['value'];
+	});
+
+	var request = {
+		url: `http://localhost:3000/users/${data.id}`,
+		method: 'PUT',
+		data: data
+	};
+
+	$.ajax(request).done(function(response) {
+		alert('Data Updated Successfully!');
+	});
 });
 
 if (window.location.pathname == '/') {
@@ -13,6 +30,25 @@ if (window.location.pathname == '/') {
 
 		var request = {
 			url: `http://localhost:3000/users/${id}`,
+			method: 'DELETE'
+		};
+
+		if (confirm('Do you really want to delete this record?')) {
+			$.ajax(request).done(function(response) {
+				alert('Data Deleted Successfully!');
+				location.reload();
+			});
+		}
+	});
+}
+
+if (window.location.pathname == '/') {
+	$ondelete = $('form a.delete-recipe');
+	$ondelete.click(function() {
+		var id = $(this).attr('data-recipe-id');
+
+		var request = {
+			url: `http://localhost:3000/recipes/${id}`,
 			method: 'DELETE'
 		};
